+++ gps/util.py
